Move Progress to design system
